+++ .storybook/preview.js
@@ -13,7 +13,7 @@ import companyListsTasks from '../src/client/components/CompanyLists/tasks/dummy
 import taskStoriesTasks from '../src/client/components/Task/__stories__/tasks.js'
 import typeaheadTasks from '../src/client/components/Typeahead2/tasks.js'
 import contactTasks from '../src/client/components/Resource/__stories__/tasks.js'
-import taskFormTasks from '../src/client/components/Task/Form/__stories__/tasks'
+import formTasks from '../src/client/components/Form/__stories__/tasks'
 import eventFormAndMetadataTasks from '../src/client/modules/Events/EventForm/__stories__/tasks'
 
 const GlobalStyle = createGlobalStyle`
@@ -42,8 +42,8 @@ addDecorator((s) => (
         ...companyListsTasks(),
         ...taskStoriesTasks,
         ...contactTasks,
-        ...taskFormTasks,
         ...typeaheadTasks,
+        ...formTasks,
         ...eventFormAndMetadataTasks,
       }}
     >
