+++ src/apps/companies/apps/edit-company/client/EditCompanyForm.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import CompanyMatched from './CompanyMatched'
 import CompanyUnmatched from './CompanyUnmatched'
-import TaskForm from '../../../../../client/components/Task/Form'
+import TaskForm from '../../../../../client/components/Form'
 import { UNITED_STATES_ID, CANADA_ID } from '../../../../../common/constants'
 import urls from '../../../../../lib/urls'
