+++ src/apps/companies/apps/referrals/send-referral/client/SendReferralForm.jsx
@@ -9,7 +9,7 @@ import LocalHeader from '../../../../../../client/components/LocalHeader/LocalHe
 import { Main } from '../../../../../../client/components/'
 import { companies, dashboard } from '../../../../../../lib/urls'
 import Task from '../../../../../../client/components/Task'
-import TaskForm from '../../../../../../client/components/Task/Form'
+import TaskForm from '../../../../../../client/components/Form'
 
 import {
   ID as STATE_ID,
