+++ src/apps/company-lists/client/CreateListForm.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { FieldInput } from '../../../client/components'
-import TaskForm from '../../../client/components/Task/Form'
+import TaskForm from '../../../client/components/Form'
 
 const CreateListForm = ({
   id,
