+++ src/apps/investments/client/opportunities/Details/CreateUKInvestmentOpportunity.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { ID, TASK_CREATE_INVESTMENT_OPPORTUNITY } from './state'
 import { FieldInput, Main } from '../../../../../client/components'
-import TaskForm from '../../../../../client/components/Task/Form'
+import TaskForm from '../../../../../client/components/Form'
 
 import urls from '../../../../../lib/urls'
